+++ CSK_Module_DigitalIOManager/scripts/Configuration/DigitalIOManager/DigitalIOManager_Model.lua
@@ -45,10 +45,12 @@ setDigitalIOManager_ModelHandle(digitalIOManager_Model)
 digitalIOManager_Model.helperFuncs = require('Configuration/DigitalIOManager/helper/funcs')
 
 digitalIOManager_Model.initialized = false -- Status if app initialized already the interfaces after boot up
+digitalIOManager_Model.trackStatus = false -- Status if app should track current status of input ports
 digitalIOManager_Model.handles = {} -- Handles of DigitalIO interfaces
 digitalIOManager_Model.portState = {} -- States of the DigitalIO interfaces
 digitalIOManager_Model.digitalInputs = {} -- List of available input interfaces
 digitalIOManager_Model.digitalOutputs = {} -- List of available input interfaces
+digitalIOManager_Model.sensorStatus = {} -- Status of sensor's measurement
 
 digitalIOManager_Model.forwardFunctions = {} -- List of functions to be used to forward incoming trigger via DigitalIn as events
 digitalIOManager_Model.triggerFunctions = {} -- List of functions to be used to forward incoming events to trigger DigitalOut
@@ -119,6 +121,7 @@ local function initialize()
   _G.logger:info(nameOfModule .. ': Initialize digital IO ports.')
   digitalIOManager_Model.parameters.active = {} -- Status if port is active to process
   digitalIOManager_Model.parameters.mode = {} -- Port is used by 'SCRIPT', 'FLOW' or 'BLOCKED'
+  digitalIOManager_Model.sensorStatus = {} -- Status of sensor's measurement
 
   digitalIOManager_Model.parameters.inDebounceValue = {} -- List of Input DebounceValues for the ports
   digitalIOManager_Model.parameters.inDebounceMode = {} -- List of Input DebounceMode for the ports
@@ -140,6 +143,7 @@ local function initialize()
     digitalIOManager_Model.parameters.inputLogic[id] = 'ACTIVE_HIGH' -- 'ACTIVE_HIGH', 'ACTIVE_LOW'
 
     digitalIOManager_Model.parameters.mode[id] = 'SCRIPT' -- 'SCRIPT', 'FLOW'
+    digitalIOManager_Model.sensorStatus[id] = '-' -- '-','true', 'false'
   end
 
   for i=1, #digitalIOManager_Model.digitalOutputs do
@@ -196,6 +200,16 @@ local function setupAll()
         digitalIOManager_Model.handles[id]:setDebounceMode(digitalIOManager_Model.parameters.inDebounceMode[id])
         digitalIOManager_Model.handles[id]:setDebounceValue(digitalIOManager_Model.parameters.inDebounceValue[id])
         digitalIOManager_Model.handles[id]:setLogic(digitalIOManager_Model.parameters.inputLogic[id])
+
+        -- Optionally track state of input signals to e.g. show them on UI
+        if digitalIOManager_Model.trackStatus then
+          Connector.DigitalIn.register(digitalIOManager_Model.handles[id], "OnChange",
+          function(state)
+            digitalIOManager_Model.sensorStatus[id]= state
+            Script.notifyEvent("DigitalIOManager_OnNewInputPortTable", digitalIOManager_Model.helperFuncs.createJsonList('input', digitalIOManager_Model.parameters.inDebounceMode, digitalIOManager_Model.parameters.active, digitalIOManager_Model.parameters.inDebounceMode, digitalIOManager_Model.parameters.inDebounceValue, digitalIOManager_Model.parameters.inputLogic, nil, digitalIOManager_Model.parameters.mode, digitalIOManager_Model.sensorStatus))
+          end)
+        end
+
         if digitalIOManager_Model.parameters.forwardEvent[id] then
           Connector.DigitalIn.register(digitalIOManager_Model.handles[id], "OnChange", digitalIOManager_Model.forwardFunctions[id])
         end